@@ -1,8 +1,10 @@
 package net.highskiesmc.fishing.events.handlers;
 
+import net.highskiesmc.fishing.HSFishing;
 import net.highskiesmc.fishing.events.events.FishCaughtEvent;
 import net.highskiesmc.fishing.util.DropEntry;
 import net.highskiesmc.fishing.util.HSFishingRod;
+import net.highskiesmc.fishing.util.HologramUtils;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
@@ -13,13 +15,58 @@ import org.bukkit.inventory.ItemStack;
 import java.text.DecimalFormat;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Random;
 import java.util.stream.Collectors;
 
 public class FishCaughtHandler implements Listener {
+    private final HSFishing MAIN;
+
+    public FishCaughtHandler(HSFishing main) {
+        this.MAIN = main;
+    }
+
     // Base message that will have its placeholders updated
     private final static String MESSAGE = "{rarity} " + ChatColor.GRAY + ChatColor.BOLD + ">> "
             + ChatColor.RED + "{amount}x {display-name} " + ChatColor.GRAY + '(' + ChatColor.WHITE + '+'
             + ChatColor.DARK_AQUA + "{xp} " + ChatColor.YELLOW + "xp" + ChatColor.GRAY + ')';
+
+
+    @EventHandler(priority = EventPriority.LOWEST)
+    public void applyExtraPerks(FishCaughtEvent e) {
+        final HSFishingRod ROD = e.getFishingRod();
+
+        List<DropEntry> drops = e.getDroppedItems();
+
+        if (ROD.getDoubleDrops() > 0) {
+            double rng = new Random().nextDouble();
+            if (rng <= ROD.getDoubleDrops()) {
+                for (DropEntry drop : drops) {
+                    drop.setAmount(drop.getAmount() * 2);
+                }
+
+                HologramUtils.spawnAnimated(this.MAIN, e.getHook().getLocation().subtract(0, 1.5, 0),
+                        ChatColor.YELLOW.toString() + ChatColor.BOLD + "DOUBLE DROPS",
+                        0, 1);
+            }
+
+        }
+
+        if (ROD.getDoubleXp() > 0) {
+            double rng = new Random().nextDouble();
+            if (rng <= ROD.getDoubleXp()) {
+                for (DropEntry drop : drops) {
+                    drop.setExperience(drop.getExperience() * 2);
+                }
+
+                HologramUtils.spawnAnimated(this.MAIN, e.getHook().getLocation().subtract(0, 1.5, 0),
+                        ChatColor.YELLOW.toString() + ChatColor.BOLD + "DOUBLE XP",
+                        0, 1);
+            }
+        }
+
+
+        e.setDroppedItems(drops);
+    }
 
     @EventHandler(priority = EventPriority.MONITOR)
     public void onFishCaught(FishCaughtEvent e) {
@@ -33,7 +80,8 @@ public class FishCaughtHandler implements Listener {
             String displayName = drop.getItemMeta().hasDisplayName()
                     ? drop.getItemMeta().getDisplayName()
                     :
-                    Arrays.stream(drop.getType().name().split("_")).map(str -> str.substring(0, 1).toUpperCase() + str.substring(1).toLowerCase()).collect(Collectors.joining(" "));
+                    Arrays.stream(drop.getType().name().split("_")).map(str -> str.substring(0,
+                            1).toUpperCase() + str.substring(1).toLowerCase()).collect(Collectors.joining(" "));
             String xp = new DecimalFormat("#.##").format(entry.getExperience() * expMulti);
             player.sendMessage(
                     MESSAGE
